feat: Lazily imported RulesFirstTimePopup
